requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
yookassa==2.3.1
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1
